+++ app/policies/account/transaction_policy.rb
@@ -1,0 +1,39 @@
+module Account
+  class TransactionPolicy < ApplicationPolicy
+    attr_reader :user, :transaction
+
+    def initialize(user, transaction)
+      super
+      @user = user
+      @transaction = transaction
+    end
+
+    def show?
+      owner?
+    end
+
+    def update?
+      owner?
+    end
+
+    def edit?
+      owner?
+    end
+
+    def destroy?
+      owner?
+    end
+
+    class Scope < Scope
+      def resolve
+        Transaction.includes(:account).where(account: { user: user })
+      end
+    end
+
+    private
+
+    def owner?
+      transaction.account_user == user
+    end
+  end
+end
