+++ app/views/account/accounts/_show_stocks.html.erb
