+++ app/views/account/accounts/new.html.erb
@@ -1,0 +1,42 @@
+<div class="title">
+  New Account
+</div>
+
+<div class="form">
+<%= form_with(url: accounts_path, scope: :account, local: true) do |form| %>
+  <% if @account.errors.any? %>
+    <div id="error_explanation">
+      <h2><%= pluralize(@account.errors.count, "error") %> prohibited this account from being saved:</h2>
+
+      <ul>
+        <% @account.errors.full_messages.each do |message| %>
+          <li><%= message %></li>
+        <% end %>
+      </ul>
+    </div>
+  <% end %>
+
+  <form>
+    <div class="form-group">
+      <%= form.label :name, class: "input-group-text" %>
+      <%= form.text_field :name, class: "form-control" %>
+    </div>
+
+    <div class="form-group">
+      <%= form.label "Savings account ?", class: "input-group-text" %>
+      <%= form.check_box :savings, class: "form-control" %>
+    </div>
+
+    <div class="form-group">
+      <%= form.label :balance, class: "input-group-text" %>
+      <%= form.text_field :balance, class: "form-control" %>
+    </div>
+
+    <div class="form-group">
+      <%= form.submit class: "btn btn-success" %>
+    </div>
+  </form>
+<% end %>
+
+  <%= link_to 'Back', accounts_path, class: "btn btn-primary" %>
+</div>
