+++ app/views/account/accounts/show.html.erb
@@ -5,6 +5,6 @@
 <% if @account.savings? %>
   <%= render 'show_savings' %>
 <% else %>
-  <%= render 'show_stocks' %>
+  <%# <%= render 'show_stocks' %> %>
   <%= render 'show_table' %>
 <% end %>
