--- app/views/accounts/_account.json.jbuilder
@@ -1,2 +1,0 @@
-json.extract! account, :id, :name, :created_at, :updated_at
-json.url account_url(account, format: :json)
