--- app/views/accounts/edit.html.erb
@@ -1,9 +1,0 @@
-<div class="title">
-  <%= @account.name %>
-</div>
-
-<div class="form">
-<%= render 'form', account: @account %>
-
-<%= link_to 'Back', accounts_path, class: "btn btn-primary" %>
-</div>
