--- app/views/accounts/index.html.erb
@@ -1,33 +1,0 @@
-<div class="title">
-  Accounts
-</div>
-
-<table class="table table-hover">
-  <thead>
-    <tr>
-      <th>Name</th>
-      <th>Balance</th>
-      <th>Date of last balance</th>
-      <th> Actions </th>
-    </tr>
-  </thead>
-
-  <tbody>
-    <% @accounts.each do |account| %>
-      <tr>
-        <td><%= link_to account.name, account_path(account), class:"account" %></td>
-        <td >$ <%= account.stock_plus_balance %></td>
-        <td ><%= account.last_balance.date.strftime('%B %d, %Y') %></td>
-        <td>
-        <%= link_to 'Month Transactions', month_transactions_account_path(account), :class => "btn btn-warning" %>
-        <%= link_to "Transactions", transactions_account_path(account), class:"btn btn-primary" %>
-        <%= link_to 'Edit', edit_account_path(account), :class => "btn btn-warning" %>
-        <%= link_to 'Delete', account, method: :delete, data: { confirm: 'Are you sure?' }, :class => "btn btn-danger" %></td>
-      </tr>
-    <% end %>
-  </tbody>
-</table>
-
-<br>
-
-<%= link_to 'New Account', new_account_path, :class => "btn btn-success ml-3" %>
