--- app/views/accounts/index.json.jbuilder
@@ -1,1 +1,0 @@
-json.array! @accounts, partial: 'accounts/account', as: :account
