--- app/views/accounts/new.html.erb
@@ -1,9 +1,0 @@
-<div class="title">
-  New Account
-</div>
-
-<div class="form">
-  <%= render 'form', account: @account %>
-
-  <%= link_to 'Back', accounts_path, class: "btn btn-primary" %>
-</div>
