--- app/views/accounts/show.json.jbuilder
@@ -1,1 +1,0 @@
-json.partial! 'accounts/account', account: @account
