+++ spec/factories/account/accounts.rb
@@ -1,5 +1,5 @@
 FactoryBot.define do
-  factory :account do
+  factory :account, class: 'Account::Account' do
     name { Faker::Company.name }
     savings { true }
     balance { Faker::Number.number(digits: 7) }
